Use long for storing unsigned 32bit integers

+++ src/freenet/client/filter/VorbisBitstreamFilter.java
@@ -60,9 +60,9 @@ public class VorbisBitstreamFilter extends OggBitstreamFilter {
 						if(magicHeader[i+1] != magicNumber[i]) invalidate();
 					}
 					//Assemble identification header
-					int vorbis_version = input.readInt();
+					long vorbis_version = input.readInt();
 					int audio_channels = input.readUnsignedByte();
-					int audio_sample_rate = input.readInt();
+					long audio_sample_rate = input.readInt();
 					int bitrate_maximum = input.readInt();
 					int bitrate_nominal = input.readInt();
 					int bitrate_minimum = input.readInt();
